キースイッチ単位書き出しをヘルパー関数化
