+++ KeySwitchManager/Sources/Runtime/Infrastructures/Storage.Xml/Cubase/MultipleCubaseWriter.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Text;
 
 using KeySwitchManager.Commons.Data;
-using KeySwitchManager.Domain.KeySwitches.Helpers;
 using KeySwitchManager.Domain.KeySwitches.Models;
 using KeySwitchManager.Infrastructures.Storage.KeySwitches.Helper;
 
@@ -29,22 +28,13 @@ namespace KeySwitchManager.Infrastructures.Storage.Xml.Cubase
 
         public void Write( IReadOnlyCollection<KeySwitch> keySwitches, IObserver<string>? loggingSubject = null )
         {
-            var multipleKeySwitches = KeySwitchHelper.GroupBy( keySwitches );
-
-            foreach( var key in multipleKeySwitches.Keys )
-            {
-                var x = multipleKeySwitches[ key ];
-                var baseDirectory = CreatePathHelper.CreateDirectoryTree( key.Item1, key.Item2, OutputDirectory );
-
-                foreach( var k in x )
-                {
-                    var filePath = CreatePathHelper.CreateFilePath( k, Suffix, baseDirectory );
-                    using var stream = filePath.OpenWriteStream();
-                    using var fileWriter = new CubaseWriter( stream, FileEncoding );
-
-                    fileWriter.Write( new[] { k }, loggingSubject );
-                }
-            }
+            MultipleWritingHelper.Write(
+                keySwitches,
+                OutputDirectory,
+                Suffix,
+                loggingSubject,
+                stream => new CubaseWriter( stream, FileEncoding )
+            );
         }
     }
 }
